moved the GetApp to test_helpers
